Updated Expense Form to add Transactions
Removed the textbox form-control of "name" for Combobox control "Category".

Set the Tab-index of form controls and the "AcceptButton" to submit form on "Enter" press.

Updated the form header text to format "Add <Income/Expense> Transaction"

Updated the success text for the Transaction to say what type of transaction it was.